Remove legacy implementation of LOBs
