+++ h2/src/main/org/h2/command/dml/Set.java
@@ -236,6 +236,7 @@ public class Set extends Prepared {
                 // just ignore the command if not starting
                 // this avoids problems when running recovery scripts
                 int value = getIntValue();
+                database.setCreateBuild(value);
                 synchronized (database) {
                     addOrUpdateSetting(name, null, value);
                 }
