+++ h2/src/main/org/h2/engine/Database.java
@@ -230,6 +230,8 @@ public class Database implements DataHandler, CastDataProvider {
 
     private Authenticator authenticator;
 
+    private int createBuild = Constants.BUILD_ID;
+
     public Database(ConnectionInfo ci, String cipher) {
         if (ASSERT) {
             META_LOCK_DEBUGGING.set(null);
@@ -534,11 +536,7 @@ public class Database implements DataHandler, CastDataProvider {
             boolean existsPage = FileUtils.exists(pageFileName);
             boolean existsMv = FileUtils.exists(mvFileName);
             if (existsData && (!existsPage && !existsMv)) {
-                throw DbException.get(
-                        ErrorCode.FILE_VERSION_ERROR_1, "Old database: " +
-                        dataFileName +
-                        " - please convert the database " +
-                        "to a SQL script and re-create it.");
+                throw getFileVersionError(dataFileName);
             }
             if (existsPage && !FileUtils.canWrite(pageFileName)) {
                 readOnly = true;
@@ -699,6 +697,12 @@ public class Database implements DataHandler, CastDataProvider {
                 }
             }
         }
+        if (!isMVStore() && createBuild < 197) {
+            // PageStore has problems due to changes in referential constraints
+            // that lead to database corruption (#1247).
+            // LOBs from 1.2.x releases are also not supported.
+            throw getFileVersionError(databaseName + Constants.SUFFIX_PAGE_FILE);
+        }
         getLobStorage().init();
         systemSession.commit(true);
 
@@ -710,6 +714,11 @@ public class Database implements DataHandler, CastDataProvider {
                 setWriteDelay(writeDelay);
             }
         }
+    }
+
+    private static DbException getFileVersionError(String dataFileName) {
+        return DbException.get(ErrorCode.FILE_VERSION_ERROR_1, "Old database: " + dataFileName
+                + " - please convert the database to a SQL script and re-create it.");
     }
 
     private void executeMeta() {
@@ -2996,6 +3005,15 @@ public class Database implements DataHandler, CastDataProvider {
     public TimeZoneProvider currentTimeZone() {
         // This method should not be reachable
         throw DbException.getUnsupportedException("Unsafe comparison or cast");
+    }
+
+    /**
+     * Sets the create build.
+     *
+     * @param createBuild the create build to set
+     */
+    public void setCreateBuild(int createBuild) {
+        this.createBuild = createBuild;
     }
 
 }
