+++ h2/src/main/org/h2/store/Data.java
@@ -46,7 +46,6 @@ import org.h2.value.ValueInterval;
 import org.h2.value.ValueJavaObject;
 import org.h2.value.ValueJson;
 import org.h2.value.ValueLob;
-import org.h2.value.ValueLobDb;
 import org.h2.value.ValueLong;
 import org.h2.value.ValueNull;
 import org.h2.value.ValueResultSet;
@@ -684,38 +683,16 @@ public class Data {
         case Value.BLOB:
         case Value.CLOB: {
             writeByte(type == Value.BLOB ? BLOB : CLOB);
-            if (v instanceof ValueLob) {
-                ValueLob lob = (ValueLob) v;
-                byte[] small = lob.getSmall();
-                if (small == null) {
-                    int t = -1;
-                    if (!lob.isLinkedToTable()) {
-                        t = -2;
-                    }
-                    writeVarInt(t);
-                    writeVarInt(lob.getTableId());
-                    writeVarInt(lob.getObjectId());
-                    writeVarLong(lob.getType().getPrecision());
-                    writeByte((byte) (lob.isCompressed() ? 1 : 0));
-                    if (t == -2) {
-                        writeString(lob.getFileName());
-                    }
-                } else {
-                    writeVarInt(small.length);
-                    write(small, 0, small.length);
-                }
+            ValueLob lob = (ValueLob) v;
+            byte[] small = lob.getSmall();
+            if (small == null) {
+                writeVarInt(-3);
+                writeVarInt(lob.getTableId());
+                writeVarLong(lob.getLobId());
+                writeVarLong(lob.getType().getPrecision());
             } else {
-                ValueLobDb lob = (ValueLobDb) v;
-                byte[] small = lob.getSmall();
-                if (small == null) {
-                    writeVarInt(-3);
-                    writeVarInt(lob.getTableId());
-                    writeVarLong(lob.getLobId());
-                    writeVarLong(lob.getType().getPrecision());
-                } else {
-                    writeVarInt(small.length);
-                    write(small, 0, small.length);
-                }
+                writeVarInt(small.length);
+                write(small, 0, small.length);
             }
             break;
         }
@@ -924,31 +901,15 @@ public class Data {
             if (smallLen >= 0) {
                 byte[] small = Utils.newBytes(smallLen);
                 read(small, 0, smallLen);
-                return ValueLobDb.createSmallLob(type == BLOB ? Value.BLOB : Value.CLOB, small);
+                return ValueLob.createSmallLob(type == BLOB ? Value.BLOB : Value.CLOB, small);
             } else if (smallLen == -3) {
                 int tableId = readVarInt();
                 long lobId = readVarLong();
                 long precision = readVarLong();
-                return ValueLobDb.create(type == BLOB ? Value.BLOB : Value.CLOB, handler, tableId,
+                return ValueLob.create(type == BLOB ? Value.BLOB : Value.CLOB, handler, tableId,
                         lobId, null, precision);
             } else {
-                int tableId = readVarInt();
-                int objectId = readVarInt();
-                long precision = 0;
-                boolean compression = false;
-                // -1: regular; -2: regular, but not linked (in this case:
-                // including file name)
-                if (smallLen == -1 || smallLen == -2) {
-                    precision = readVarLong();
-                    compression = readByte() == 1;
-                }
-                if (smallLen == -2) {
-                    String filename = readString();
-                    return ValueLob.openUnlinked(type == BLOB ? Value.BLOB : Value.CLOB, handler, tableId,
-                            objectId, precision, compression, filename);
-                }
-                return ValueLob.openLinked(type == BLOB ? Value.BLOB : Value.CLOB, handler, tableId,
-                        objectId, precision, compression);
+                throw DbException.get(ErrorCode.FILE_CORRUPTED_1, "lob type: " + smallLen);
             }
         }
         case ARRAY:
@@ -1173,38 +1134,16 @@ public class Data {
         case Value.BLOB:
         case Value.CLOB: {
             int len = 1;
-            if (v instanceof ValueLob) {
-                ValueLob lob = (ValueLob) v;
-                byte[] small = lob.getSmall();
-                if (small == null) {
-                    int t = -1;
-                    if (!lob.isLinkedToTable()) {
-                        t = -2;
-                    }
-                    len += getVarIntLen(t);
-                    len += getVarIntLen(lob.getTableId());
-                    len += getVarIntLen(lob.getObjectId());
-                    len += getVarLongLen(lob.getType().getPrecision());
-                    len += 1;
-                    if (t == -2) {
-                        len += getStringLen(lob.getFileName());
-                    }
-                } else {
-                    len += getVarIntLen(small.length);
-                    len += small.length;
-                }
+            ValueLob lob = (ValueLob) v;
+            byte[] small = lob.getSmall();
+            if (small == null) {
+                len += getVarIntLen(-3);
+                len += getVarIntLen(lob.getTableId());
+                len += getVarLongLen(lob.getLobId());
+                len += getVarLongLen(lob.getType().getPrecision());
             } else {
-                ValueLobDb lob = (ValueLobDb) v;
-                byte[] small = lob.getSmall();
-                if (small == null) {
-                    len += getVarIntLen(-3);
-                    len += getVarIntLen(lob.getTableId());
-                    len += getVarLongLen(lob.getLobId());
-                    len += getVarLongLen(lob.getType().getPrecision());
-                } else {
-                    len += getVarIntLen(small.length);
-                    len += small.length;
-                }
+                len += getVarIntLen(small.length);
+                len += small.length;
             }
             return len;
         }
