+++ h2/src/main/org/h2/store/LobStorageFrontend.java
@@ -10,7 +10,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.Reader;
 import org.h2.value.Value;
-import org.h2.value.ValueLobDb;
+import org.h2.value.ValueLob;
 
 /**
  * This factory creates in-memory objects and temporary files. It is used on the
@@ -40,7 +40,7 @@ public class LobStorageFrontend implements LobStorageInterface {
     }
 
     @Override
-    public void removeLob(ValueLobDb lob) {
+    public void removeLob(ValueLob lob) {
         // not stored in the database
     }
 
@@ -53,7 +53,7 @@ public class LobStorageFrontend implements LobStorageInterface {
      * @return the stream
      */
     @Override
-    public InputStream getInputStream(ValueLobDb lob, byte[] hmac,
+    public InputStream getInputStream(ValueLob lob, byte[] hmac,
             long byteCount) throws IOException {
         if (byteCount < 0) {
             byteCount = Long.MAX_VALUE;
@@ -68,7 +68,7 @@ public class LobStorageFrontend implements LobStorageInterface {
     }
 
     @Override
-    public ValueLobDb copyLob(ValueLobDb old, int tableId, long length) {
+    public ValueLob copyLob(ValueLob old, int tableId, long length) {
         throw new UnsupportedOperationException();
     }
 
@@ -82,7 +82,7 @@ public class LobStorageFrontend implements LobStorageInterface {
         // need to use a temp file, because the input stream could come from
         // the same database, which would create a weird situation (trying
         // to read a block while writing something)
-        return ValueLobDb.createTempBlob(in, maxLength, handler);
+        return ValueLob.createTempBlob(in, maxLength, handler);
     }
 
     /**
@@ -97,7 +97,7 @@ public class LobStorageFrontend implements LobStorageInterface {
         // need to use a temp file, because the input stream could come from
         // the same database, which would create a weird situation (trying
         // to read a block while writing something)
-        return ValueLobDb.createTempClob(reader, maxLength, handler);
+        return ValueLob.createTempClob(reader, maxLength, handler);
     }
 
     @Override
